dj_backend_server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
dataclasses-json==0.5.14
Django==4.2.3
dnspython==2.4.1
exceptiongroup==1.1.2
frozenlist==1.4.0
grpcio==1.56.2
grpcio-tools==1.56.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
kombu==5.3.1
langchain==0.0.247
langsmith==0.0.15
loguru==0.7.0
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==10.0.0
pinecone-client==2.2.2
portalocker==2.7.0
prompt-toolkit==3.0.39
protobuf==4.23.4
pycparser==2.21
pydantic==1.10.12
PyMySQL==1.1.0
pypdf==3.14.0
pypdfium2==4.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
qdrant-client==1.3.2
redis==4.6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
django-cors-headers==4.2.0